+++ ui_automation_tests/step_defs/test_edit_open_application.py
@@ -20,3 +20,9 @@ def i_remove_a_good_from_the_application(driver):
 @then("no goods types are left on the application")
 def no_goods_types_are_left_on_the_application(driver):
     assert (ApplicationGoodsTypeList(driver).find_remove_goods_type_link(), None)
+
+
+@when("I click on open goods tile")  # noqa
+def click_my_goods_link(driver):  # noqa
+    exporter_hub = ApplicationOverviewPage(driver)
+    exporter_hub.click_open_goods_link()
