+++ ui_automation_tests/step_defs/test_edit_standard_application.py
@@ -1,16 +1,10 @@
 from pytest_bdd import when, scenarios, then
 
-from shared import functions
 from conftest import enter_application_name, enter_export_licence
 from pages.application_overview_page import ApplicationOverviewPage
 from pages.application_page import ApplicationPage
 
 scenarios("../features/edit_standard_application.feature", strict_gherkin=False)
-
-
-@when("I click back to the application overview")
-def i_click_on_application_overview(driver):
-    functions.click_back_link(driver)
 
 
 @when("I click on the application third parties link")
