+++ ui_automation_tests/step_defs/test_search_and_filter_goods.py
@@ -77,3 +77,8 @@ def good_control_code_is_found(driver, control_code):
 def total_goods_found(driver, total):
     total_goods = len(ApplicationGoodsList(driver).get_good_descriptions())
     assert total_goods == int(total), "Incorrect number of goods matching search criteria were found"
+
+
+@when("I click the add from organisations goods button")  # noqa
+def click_add_from_organisation_button(driver):  # noqa
+    driver.find_element_by_css_selector('a[href*="add-preexisting"]').click()
