+++ ui_automation_tests/step_defs/test_submit_open_application.py
@@ -135,3 +135,9 @@ def see_all_or_no_selected(driver, assigned_or_unassigned):
         assert countries_page.all_selected()
     else:
         assert countries_page.all_deselected()
+
+
+@when("I click on the goods link from overview")  # noqa
+def click_goods_link_overview(driver):  # noqa
+    overview_page = ApplicationOverviewPage(driver)
+    overview_page.click_open_goods_link()
